[MRESOLVER-248] Make BF and DF collectors coexists
